Add hook for a callback after copying to clipboard
The primary intended purpose of this functionality is to allow wrapper
utilities such as passmenu to do something after pass has copied the
password to the clipboard. This is not previously possible since xclip
prevents the pass script from exiting.

How to use: export a function named PASSWORD_STORE_POST_COPY_HOOK. If
present, this function will be called with two arguments:

 1. The name of the copied password (`"foo/bar"` when pass is invoked as
    `"pass foo/bar"`)
 2. A copy of the message pass prints to the console

Example:

    PASSWORD_STORE_POST_COPY_HOOK() {
      notify-send -u normal -t 2000 "Copied $1" "$2"
    }
    export -f PASSWORD_STORE_POST_COPY_HOOK
    pass foo/bar